Fix non-imported IAM roles and clear commented backwards compatibility snippets
